Read scores from 'teams' field instead of calculating from 'scoringPlays'
Once in a full moon, the 'scoringPlays' field seems to remain empty for a
long time even when goals have been scored. Read the team scores from the
'teams' field which hopefully updates more reliably. This also makes it
possible to detect discrepancies between the scores and goals.

The scores could also be read from the 'linescore' field but that field is
missing from older test data files, and it would be quite a bit of work to
retrieve it realistically.

(Hopefully, partially) solves #3.