Parse playoff series wins from series summary description again
I changed the source in a5ec4a69cfed89d66cbfc112c6160c1deabff8e4 to
`matchup-teams` because it seemed to update quicker than
`series-summary-description`. However, currently the numbers in
`matchup-teams` are missing the last two (!) games, whereas the series
summary is up-to-date. So change the source back to the previous one.