Unify ApiRequest implementations
Implement the archival logic in the archive-with-context? method, and call
it from the archive? convenience method. The former is always used in app
code and the latter in some tests, and this change ensures we execute the
same code in the app and tests (i.e. archive-with-context?).

StandingsApiRequest was already implemented this way, so no changes needed
there.